+++ src/TinyCsv/CsvOptions.cs
@@ -31,12 +31,12 @@ namespace TinyCsv
 {
     using System;
     using System.Collections.Concurrent;
-    using System.Collections.Generic;
     using System.Linq;
     using System.Reflection;
     using System.Text;
     using TinyCsv.Args;
     using TinyCsv.Attributes;
+    using TinyCsv.Helpers;
 
     /// <summary>
     ///  Csv Options abstract definition
@@ -225,7 +225,7 @@ namespace TinyCsv
         /// <summary>
         /// Object's faster properties
         /// </summary>
-        internal ConcurrentDictionary<string, Extensions.StringArrayExtensions.PropertyHelper<T>> FasterProperties { get; set; }
+        internal ConcurrentDictionary<string, PropertyHelper<T>> FasterProperties { get; set; }
 
         /// <summary>
         /// Create a CsvOptions
@@ -255,7 +255,7 @@ namespace TinyCsv
         /// <param name="type"></param>
         private void SetOptionPropertiesFromType(Type type)
         {
-            FasterProperties = Extensions.StringArrayExtensions.PropertyHelper<T>.GetPropertiesExpressionTree(type);
+            FasterProperties = PropertyHelper<T>.GetPropertiesExpressionTree(type);
             Properties = new ConcurrentDictionary<string, PropertyInfo>();
             var properties = type.GetProperties(BindingFlags.Instance | BindingFlags.Public);
             foreach (var property in properties)
