+++ src/TinyCsv/Extensions/GenericTypeExtensions.cs
@@ -1,6 +1,4 @@
-﻿
-using System.Linq;
-/// <summary>
+﻿/// <summary>
 /// 
 /// The MIT License (MIT)
 /// 
@@ -30,6 +28,8 @@ using System.Linq;
 /// </summary>
 namespace TinyCsv.Extensions
 {
+    using System.Linq;
+
     /// <summary>
     /// Generic Type extensions
     /// </summary>
@@ -47,13 +47,8 @@ namespace TinyCsv.Extensions
             var values = options.Columns.Select(column =>
             {
                 var propertyName = column.ColumnNameInternal;
-
-                //var property = options.Properties[propertyName];
-                //var value = property.GetValue(model);
-
                 var property = options.FasterProperties[propertyName];
                 var value = property.Getter(model);
-
                 var stringValue = column.Converter.Convert(value, null, column.ColumnFormatProvider);
                 return stringValue?.EnclosedInQuotesIfNecessary(options);
             });
